+++ android/src/main/java/im/zego/zim_flutter/internal/ZIMPluginConverter.java
@@ -509,7 +509,7 @@ public class ZIMPluginConverter {
     static public HashMap<String,Object>mZIMMessageReceiptInfo(ZIMMessageReceiptInfo info){
         HashMap<String,Object> infoModel = new HashMap<>();
         infoModel.put("conversationID",info.conversationID);
-        infoModel.put("conversationType",info.conversationType);
+        infoModel.put("conversationType",info.conversationType.value());
         infoModel.put("messageID",info.messageID);
         infoModel.put("status",info.status.value());
         infoModel.put("readMemberCount",info.readMemberCount);
@@ -795,7 +795,7 @@ public class ZIMPluginConverter {
         groupMemberInfoMap.put("memberRole",groupMemberInfo.memberRole);
         groupMemberInfoMap.put("userID",groupMemberInfo.userID);
         groupMemberInfoMap.put("userName",groupMemberInfo.userName);
-        groupMemberInfoMap.put("memberAvatarUrl",groupMemberInfo.memberAvatarUrl);
+        groupMemberInfoMap.put("memberAvatarUrl",groupMemberInfo.memberAvatarUrl != null?groupMemberInfo.memberAvatarUrl:"");
         return groupMemberInfoMap;
     }