+++ lib/src/internal/zim_converter.dart
@@ -227,7 +227,7 @@ class ZIMConverter {
         message ??= ZIMBarrageMessage(message: resultMap['message']);
         break;
       case ZIMMessageType.image:
-        message ??= ZIMImageMessage(resultMap['fileLocalPath']);
+        message ??= ZIMImageMessage(resultMap['fileLocalPath'] ?? "");
         message as ZIMImageMessage;
         message.thumbnailDownloadUrl = resultMap['thumbnailDownloadUrl'];
         message.thumbnailLocalPath = resultMap['thumbnailLocalPath'];
@@ -241,16 +241,16 @@ class ZIMConverter {
         message.thumbnailWidth = resultMap['thumbnailWidth'];
         break;
       case ZIMMessageType.file:
-        message ??= ZIMFileMessage(resultMap['fileLocalPath']);
+        message ??= ZIMFileMessage(resultMap['fileLocalPath'] ?? "");
         message as ZIMFileMessage;
         break;
       case ZIMMessageType.audio:
-        message ??= ZIMAudioMessage(resultMap['fileLocalPath']);
+        message ??= ZIMAudioMessage(resultMap['fileLocalPath'] ?? "");
         message as ZIMAudioMessage;
         message.audioDuration = resultMap['audioDuration'];
         break;
       case ZIMMessageType.video:
-        message ??= ZIMVideoMessage(resultMap['fileLocalPath']);
+        message ??= ZIMVideoMessage(resultMap['fileLocalPath'] ?? "");
         message as ZIMVideoMessage;
         message.videoDuration = resultMap['videoDuration'];
         message.videoFirstFrameDownloadUrl =
@@ -877,7 +877,7 @@ class ZIMConverter {
   static Map mZIMCallInviteConfig(ZIMCallInviteConfig config) {
     Map configMap = {};
     configMap['timeout'] = config.timeout;
-    configMap['payload'] = config.payload;
+    configMap['extendedData'] = config.extendedData;
     configMap['pushConfig'] = mZIMPushConfig(config.pushConfig);
     return configMap;
   }
@@ -1090,15 +1090,15 @@ class ZIMConverter {
       oZIMConversationMessageReceiptReadSentResult(Map resultMap) {
     return ZIMConversationMessageReceiptReadSentResult(
         conversationID: resultMap['conversationID'],
-        conversationType: resultMap['conversationType']);
+        conversationType: ZIMConversationTypeExtension.mapValue[resultMap['conversationType']]!);
   }
 
   static ZIMMessageReceiptsReadSentResult oZIMMessageReceiptReadSentResult(
       Map resultMap) {
     return ZIMMessageReceiptsReadSentResult(
         conversationID: resultMap['conversationID'],
-        conversationType: resultMap['conversationType'],
-        errorMessageIDs: resultMap['errorMessageIDs']);
+        conversationType: ZIMConversationTypeExtension.mapValue[resultMap['conversationType']]!,
+        errorMessageIDs: List.from(resultMap['errorMessageIDs']));
   }
 
   static ZIMMessageReceiptsInfoQueriedResult
@@ -1110,13 +1110,13 @@ class ZIMConverter {
       infos.add(oZIMMessageReceiptInfo(info));
     }
     return ZIMMessageReceiptsInfoQueriedResult(
-        errorMessageIDs: resultMap['errorMessageIDs'], infos: infos);
+        errorMessageIDs: List.from(resultMap['errorMessageIDs']), infos: infos);
   }
 
   static ZIMMessageReceiptInfo oZIMMessageReceiptInfo(Map infoMap) {
     ZIMMessageReceiptInfo receiptInfo = ZIMMessageReceiptInfo(
         conversationID: infoMap['conversationID'],
-        conversationType: infoMap['conversationType'],
+        conversationType: ZIMConversationTypeExtension.mapValue[infoMap['conversationType']]!,
         messageID: infoMap['messageID'],
         status: ZIMMessageReceiptStatusExtension.mapValue[infoMap['status']]!,
         readMemberCount: infoMap['readMemberCount'],
@@ -1132,7 +1132,7 @@ class ZIMConverter {
 
     return ZIMGroupMessageReceiptMemberListQueriedResult(
         groupID: resultMap['groupID'],
-        nextFlag: resultMap['nexFlag'],
+        nextFlag: resultMap['nextFlag'],
         userList: userList);
   }