+++ lib/src/internal/zim_manager.dart
@@ -8,7 +8,7 @@ import 'package:zego_zim/zego_zim.dart';
 
 class ZIMManager {
   static const MethodChannel channel = MethodChannel('zego_zim_plugin');
-  static StreamSubscription<dynamic>? streamSubscription;
+//  static StreamSubscription<dynamic>? streamSubscription;
 
   static int handleSequence = 0;